(#684) conversion from ajv to json-schema-library and update to typescript 5.5.3
